Rename and restructure view
