+++ src/pages/Auth/index.ts
@@ -1,0 +1,4 @@
+export * from './ForgotPasswordPage';
+export * from './LoginPage';
+export * from './SignUpPage';
+export * from './OnboardingPage';
