+++ src/pages/Authenticated/index.ts
@@ -1,1 +1,1 @@
-export * from './AvailableFacilities';
+export * from './AvailableFacilitiesPage';
