+++ src/pages/LandingPage.tsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export const LandingPage: React.FC = () => {
+  return <>Layout</>;
+};
