--- src/pages/Public/index.ts
@@ -1,5 +1,0 @@
-export * from './ForgotPassword';
-export * from './Login';
-export * from './SignUp';
-export * from './Onboarding';
-export * from './404';
