+++ src/pages/index.ts
@@ -1,2 +1,4 @@
-export * from './Public';
+export * from './Auth';
 export * from './Authenticated';
+export * from './404';
+export * from './LandingPage';
